Assess one operator at a time
This way you don't evaluate more than one pair of numbers at a time
Assess one operator at a time to mimic the following behavior:

Enter a number (12).

Enter an operator (+).

Enter a second number (7).

Enter a second operator (-). At this point, it should evaluate the
initial pair of numbers (12 + 7), then display the result (19).

Enter another number (1).

Enter another operator or equals sign (=). At this point, it should
use the previous result (19) as the first number, the operator (-), and
the new number (1) to calculate the new equation 19 - 1. You should see
the result (18) on the display.